simulation of a propagating sinusoidal striking a dielectric medium

+++ fd1d/cuda/fd1d_1_4.cu
@@ -15,8 +15,8 @@
 
 __device__
 float sinusoidal(int t, float ds, float freq) {
-    float dt = ds/6e8;  /* time step (s) */
-    return sin(2*M_PI*freq*dt*t);
+    float dt = ds/6e8f;  /* time step (s) */
+    return sinf(2*M_PI*freq*dt*t);
 }
 
 
@@ -26,7 +26,7 @@ void exfield(int t, int nx, float *cb, float *ex, float *hy) {
     for (int i = idx+1; i < nx; i += stx)
         ex[i] += cb[i] * (hy[i-1] - hy[i]);
     /* put a sinusoidal wave at the low end */
-    if (idx == 1) ex[1] += sinusoidal(t, 0.01f, 700e6);
+    if (idx == 1) ex[1] += sinusoidal(t, 0.01f, 700e6f);
 }
 
 
@@ -37,7 +37,7 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     if (idx == nx-1) ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = idx; i < nx-1; i += stx)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
 
 
@@ -45,7 +45,7 @@ float *dielectric(int nx, float epsr) {
     float *cb;
     cudaMallocManaged(&cb, nx*sizeof(*cb));
     for (int i = 0; i < nx; cb[i] = 0.5f, i++);
-    for (int i = nx/2; i < nx; cb[i] = 0.5/epsr, i++);
+    for (int i = nx/2; i < nx; cb[i] = 0.5f/epsr, i++);
     return cb;
 }
 
@@ -70,9 +70,9 @@ int main() {
     cudaMallocManaged(&bc, 4*sizeof(*bc));
     for (int i = 0; i < 4; bc[i] = 0.0f, i++);
 
-    float ds = 0.01;  /* spatial step (m) */
-    float dt = ds/6e8;  /* time step (s) */
-    float epsr = 4.0;  /* relative permittivity */
+    float ds = 0.01f;  /* spatial step (m) */
+    float dt = ds/6e8f;  /* time step (s) */
+    float epsr = 4.0f;  /* relative permittivity */
     float *cb = dielectric(nx, epsr);
 
     int numSM;
